Do not consider testFixtures* a runtime variant
The java-test-fixtures plugin creates configurations that meet criteria
for a runtime variant, but also include a dependency on the current
project. This results in the manifest declaring an optional dependency
on itself.

This fix will exclude any configuration that starts with testFixtures.
It seems like a reasonable compromise since testFixtures are intended
for testing and not for runtime dependencies.

Fixes #175