Populate Plugin-Version
This respects the current behavior of appending a timestamp to versions
that end in -SNAPSHOT, but it means that the task is frequently
out-of-date. Prior to 1.0.0, it should be determined whether this plugin
providing partial version calculation is beneficial, given that it means
all downstream tasks are always invalidated and many other plugins exist
for gradle versioning.

JENKINS-58205