Introduce generatedJenkinsTest
Generate the PluginAutomaticTestBuilder test class with JavaPoet and
ensure it compiles without warnings. Add coverage to ensure these
generated tests pass and fail when expected.

The generation task requiring JavaPoet is run through the worker api,
which allows for keeping the library and the jenkins-test-harness off
the buildscript classpath. The generation task is still disabled by
default.

Instead of adding the generated test to the built-in test task, create
its own generatedJenkinsTest task. This makes it easy to run the
built-in tests with JUnit 5 or TestNG while still running the generated
tests with JUnit 4. The separate task is able to run in parallel with
other test tasks by default. Check depends on this new test task.

Add coverage to ensure JenkinsRule tests work out of the box without
warnings for the test sourceSet. This required adding the
jcip-annotations and findbugs to testCompileOnly configuration.

Issue #180