Specify path to war by system property
jenkins-test-harness will prefer a system property for unpacking the war
if present[1]. If it is not present, it will attempt to find it on the
classpath.

This change introduces a single-file configuration that stores the
declared version of the jenkins war. This value is then passed as a
system property to the test and generatedJenkinsTest tasks. This allows
us to remove jenkins-war and its transitive dependencies from the
classpath.

[1]: https://github.com/jenkinsci/jenkins-test-harness/blob/jenkins-test-harness-2.71/src/main/java/org/jvnet/hudson/test/WarExploder.java#L94