Separate buildDirectory settings for test tasks
The buildDirectory system property is used for unpacking the Jenkins war
file[1] and installing plugins into. GeneratedJenkinsTests only need the
plugins on the runtimeClasspath, but test also packages plugins on the
testRuntimeClasspath.

Relocate these directories under the build/jpi-plugin folder so it's
clearer the jpi plugin produces these files.

[1]: https://github.com/jenkinsci/jenkins-test-harness/blob/jenkins-test-harness-2.71/src/main/java/org/jvnet/hudson/test/WarExploder.java#L125
Related to #180