feat: Support Jenkins >= 2.475
Jenkins 2.475 and Jenkins LTS 2.479.1 introduce a change to the Java Servlet API version and the servlet container support policy.

This takes it from JEE8 to JEE9.
This changes the coordinates of servlet-api to jakarta.

This gradle plugin can be built with JDK 11 but not with JDK 17 which would be required for adding the extra test to `JenkinsVersionIntegrationSpec`.

Related: JENKINS-73278